Update bulk amounts of repositories in parallel
I have a script `git-update-subdir-repos` which performs the following
commands on every sub-directory:

1. `git fetch origin`
2. `git merge --ff-only origin/master`

The script uses `fd ... --exec git ...` to run the commands.
The problem with this approach, though, is that the script runs
`fd` twice.  It is a waste of time to iterate through the list of
directories once per each command described above.  Obviously it would
be more efficient to execute both Git commands together with only one
invocation of `fd` so that we build our list of directories only once.

This patch makes the `git-update-subdir-repos` script more efficient by
doing exactly that: we now call `fd` only once.  Instead of combining
the commands within our `fd` invocation, i.e. with its `--exec` option,
we now pipe the output into GNU Parallel.  Not only does GNU Parallel
allow us to easily combine the commands, it also obviously lets the
system run these commands in parallel to improve efficiency.

Finally, the patch modifies the README to list GNU Parallel as one
of the requirements of some scripts, with this script currently being
the only such example.