Add a script to clone a repo then eval a command on it
The script contains a detailed example, so please read that for
information about both how this script works and why it can be useful
in certain scenarios.

While the script itself mentions this, I feel it is worth repeating
here that this script works **only** with the Fish shell.  It should
not be hard to port this to GNU Bash or even `/bin/sh` though, and
of course patches are always welcome.